post/requirements.txt
beautifulsoup4==4.10.0
boto3==1.20.13
botocore==1.23.14
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
jmespath==0.10.0
praw==7.6.0
prawcore==2.3.0
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.3.1
update-checker==0.18.0
urllib3==1.26.7
websocket-client==1.2.1
python-dotenv==0.19.0
